added weather util to create weather result object

+++ src/main/java/edu/depaul/springbootweatherpoc/weather/model/Location.java
@@ -4,8 +4,8 @@ public class Location {
 
     private String cityName;
     private int zipCode;
-    private float latitude;
-    private float longitude;
+    private Double latitude;
+    private Double longitude;
 
     public String getCityName() {
         return cityName;
@@ -23,19 +23,19 @@ public class Location {
         this.zipCode = zipCode;
     }
 
-    public float getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(float latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
 
-    public float getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(float longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
 }
